Move the crossfading player to a plugin
Move the crossfading player logic to a new plugin.

* Add `src/plugins/crossfadingPlayer/plugin.js` to contain the crossfading logic, including functions for setting crossfade duration, hijacking media elements, detecting silence, and initializing the plugin.
* Remove crossfading logic from `src/components/audioEngine/crossfader.logic.ts` and import the new plugin functions.
* Integrate the new crossfading player plugin with the master audio output in `src/components/audioEngine/master.logic.ts`.
* Integrate the new crossfading player plugin with the HTML audio player plugin in `src/plugins/htmlAudioPlayer/plugin.js`.
* Add user-configurable settings for gapless playback and silence detection in `src/scripts/settings/userSettings.js`.
* Add user interface elements for adjusting crossfading settings, gapless playback, and silence detection in `src/components/playbackSettings/playbackSettings.js`.
* Add a crossfade slider next to the volume slider in the now playing page in `src/controllers/playback/queue/index.html`.
* Add `src/plugins/crossfadingPlayer/plugin.ts` to convert the plugin to TypeScript and add TypeScript types for the crossfading logic.
