+++ src/controllers/playback/queue/index.html
@@ -101,6 +101,14 @@
 
                     </div>
                 </div>
+                <div class="sliderContainer flex" dir="ltr" id="crossfadeSlider">
+                    <div class="crossfadeTime"></div>
+                    <div class="nowPlayingCrossfadeSliderContainer">
+                        <input type="range" is="emby-slider" pin step="0.01" min="0" max="10" value="0"
+                            class="nowPlayingCrossfadeSlider" data-slider-keep-progress="true" />
+                    </div>
+                    <div class="crossfadeDuration"></div>
+                </div>
             </div>
         </div>
         <div class="remoteControlSection">
